Remove older notification model;
handling of this method of notification subscription isn't needed if
the fanning out is handled by redis.
