+++ app/Models/Chat/Channel.php
@@ -5,10 +5,8 @@
 
 namespace App\Models\Chat;
 
-use App\Events\UserSubscriptionChangeEvent;
 use App\Exceptions\API;
 use App\Models\Multiplayer\Match;
-use App\Models\Notification;
 use App\Models\User;
 use Carbon\Carbon;
 use ChaseConey\LaravelDatadogHelper\Datadog;
@@ -200,7 +198,6 @@ class Channel extends Model
 
         if ($this->isPM()) {
             $this->unhide();
-            broadcast_notification(Notification::CHANNEL_MESSAGE, $message, $sender);
         }
 
         Datadog::increment('chat.channel.send', 1, ['target' => $this->type]);
@@ -224,10 +221,6 @@ class Channel extends Model
             $userChannel->save();
         }
 
-        if ($this->isPM()) {
-            event(new UserSubscriptionChangeEvent('add', $user, $this));
-        }
-
         Datadog::increment('chat.channel.join', 1, ['type' => $this->type]);
     }
 
@@ -244,7 +237,6 @@ class Channel extends Model
         }
 
         if ($this->isPM()) {
-            event(new UserSubscriptionChangeEvent('remove', $user, $this));
             $userChannel->update(['hidden' => true]);
         } else {
             $userChannel->delete();
@@ -268,16 +260,10 @@ class Channel extends Model
             return;
         }
 
-        $hiddenUserChannels = UserChannel::where([
+        UserChannel::where([
             'channel_id' => $this->channel_id,
             'hidden' => true,
-        ]);
-
-        foreach ($hiddenUserChannels->get() as $userChannel) {
-            event(new UserSubscriptionChangeEvent('add', $userChannel->user, $this));
-        }
-
-        $hiddenUserChannels->update([
+        ])->update([
             'hidden' => false,
         ]);
     }
